Crate som_gc

Modules

active_plan

api

collection

gc_interface

gcref

gcslice

object_model

reference_glue

scanning

slot

Macros

debug_assert_valid_semispace_ptr

debug_assert_valid_semispace_ptr_value

Structs

SOMVM

Statics

MMTK_SINGLETON

Functions

handshake_with_vm

Initialize globals to allow VM/MMTk communication. TODO: Ideally, this would also take a closure that initializes the VM-specific globals, namely the UNIVERSE_RAW_POINTER needed for the VM to report roots given the universe when MMTk requests them from the VM. As such, this function only contains MOST of the handshake logic.