som_gc

Function handshake_with_vm

pub fn handshake_with_vm(
    gc_interface_ptr: &mut Box<GCInterface>,
    vm_callbacks: MMTKtoVMCallbacks,
)

Initialize globals to allow VM/MMTk communication. TODO: Ideally, this would also take a closure that initializes the VM-specific globals, namely the UNIVERSE_RAW_POINTER needed for the VM to report roots given the universe when MMTk requests them from the VM. As such, this function only contains MOST of the handshake logic.