som_gc

Module gc_interface

Structs

GCInterface

Interface from VM to the GC logic (MMTk)

MMTKtoVMCallbacks

Callbacks used to provide MMTk->VM communication.

Enums

AllocSiteMarker

Explicitly mentions what an allocation was requested for. The intent is to help debugging: we can track where GC was triggered, and what triggered it. This is all to find GC bugs, and also track memory usage for experiments.

Constants

BIGINT_MAGIC_ID

STRING_MAGIC_ID

Statics

WORLD_LOCK

Traits

GcType

Implements a per-type magic number. GC needs to access type info from raw ObjectReference types, so data that gets put on the GC heap has an associated type ID that gets put in a per-allocation header.

SOMAllocator

All functions necessary to allocate memory from within som-rs.

SliceConstraint

To save on some space in the trait itself. Bit overkill, probably