som_gc::gc_interface

Trait HasTypeInfoForGC

pub trait HasTypeInfoForGC {
    // Required method
    fn get_magic_gc_id() -> u8;
}

Implements a per-type magic number. GC needs to access type info from raw ObjectReference types, so data that gets put on the GC heap has an associated type ID that gets put in a per-allocation header.

Required Methods

fn get_magic_gc_id() -> u8

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.